Reorganize curation subapp
