+++ src/indra_cogex/apps/curator/__init__.py
@@ -1,1 +1,7 @@
 """Curation app for INDRA CoGEx."""
+
+from .curator_blueprint import curator_blueprint
+
+__all__ = [
+    "curator_blueprint",
+]
